backend/requirements.txt
fastapi==0.115.12
httpx==0.28.1
psycopg[binary]==3.2.9
python-dotenv==1.0.1
SQLAlchemy==2.0.40
uvicorn==0.34.2
